Bug fix: leveldb:read_range/5
Fixes the read_range function's result when all range is read but still returning result with a contiuation instead of 'complete' atom because of chunk size limit is reached.

This fix checks the next key to see it range is completely read.
Related eunit test functions added to test/leveldb_test.erl