Add support to merge and sort ascending order
merge_sorted_kvls/1 is now merge_sorted_kvls/2 where first arg is direction and second is kvls.
Comparison of keys in merge_sorted_kvls/2 are changed. Now if keys are the same, then the key that is provided in preceding kvl is preceding in merged kvl also.
Thiscomparison also applies to sort_kvl/2 function.